+++ pkg/analysis_server/test/src/utilities/flutter_test.dart
@@ -444,19 +444,19 @@ class MyContainer extends Container {}
 class NotFlutter {}
 class NotWidget extends State {}
 ''');
-    var myStatelessWidget = testUnitElement.getType('MyStatelessWidget');
+    var myStatelessWidget = testUnitElement.getClass('MyStatelessWidget');
     expect(_flutter.isWidget(myStatelessWidget), isTrue);
 
-    var myStatefulWidget = testUnitElement.getType('MyStatefulWidget');
+    var myStatefulWidget = testUnitElement.getClass('MyStatefulWidget');
     expect(_flutter.isWidget(myStatefulWidget), isTrue);
 
-    var myContainer = testUnitElement.getType('MyContainer');
+    var myContainer = testUnitElement.getClass('MyContainer');
     expect(_flutter.isWidget(myContainer), isTrue);
 
-    var notFlutter = testUnitElement.getType('NotFlutter');
+    var notFlutter = testUnitElement.getClass('NotFlutter');
     expect(_flutter.isWidget(notFlutter), isFalse);
 
-    var notWidget = testUnitElement.getType('NotWidget');
+    var notWidget = testUnitElement.getClass('NotWidget');
     expect(_flutter.isWidget(notWidget), isFalse);
   }