+++ pkg/analyzer/test/src/diagnostics/mixin_super_class_constraint_deferred_class_test.dart
@@ -25,7 +25,7 @@ mixin M on math.Random {}
           48, 11),
     ]);
     var mathImport = findElement.import('dart:math');
-    var randomElement = mathImport.importedLibrary!.getType('Random')!;
+    var randomElement = mathImport.importedLibrary!.getClass('Random')!;
 
     var element = findElement.mixin('M');
     assertElementTypes(element.superclassConstraints, ['Random']);