+++ pkg/analyzer/test/src/task/strong/inferred_type_test.dart
@@ -2233,7 +2233,7 @@ class D<T> {
 }
 typedef void F<V>(V v);
 ''');
-    var f = _resultUnitElement.getType('C')!.methods[0];
+    var f = _resultUnitElement.getClass('C')!.methods[0];
     _assertTypeStr(f.type, 'void Function(U) Function<U>(U)');
   }
 
@@ -2247,7 +2247,7 @@ abstract class D<T> {
 }
 typedef List<V> G<V>();
 ''');
-    var f = _resultUnitElement.getType('C')!.methods[0];
+    var f = _resultUnitElement.getClass('C')!.methods[0];
     _assertTypeStr(f.type, 'void Function<U>(List<U> Function())');
   }
 
@@ -2261,7 +2261,7 @@ class D<T> {
 }
 typedef V F<V>();
 ''');
-    var f = _resultUnitElement.getType('C')!.methods[0];
+    var f = _resultUnitElement.getClass('C')!.methods[0];
     _assertTypeStr(f.type, 'U Function() Function<U>(U)');
   }
 
@@ -2974,7 +2974,7 @@ class D {
   int foo;
 }
 ''');
-    var f = _resultUnitElement.getType('C')!.accessors[0];
+    var f = _resultUnitElement.getClass('C')!.accessors[0];
     _assertTypeStr(f.type, 'void Function(int)');
   }
 
@@ -2987,7 +2987,7 @@ class D {
   set foo(int x) {}
 }
 ''');
-    var f = _resultUnitElement.getType('C')!.accessors[0];
+    var f = _resultUnitElement.getClass('C')!.accessors[0];
     _assertTypeStr(f.type, 'void Function(int)');
   }
 
@@ -3001,10 +3001,10 @@ class C {
   };
 }
 ''');
-    var x = _resultUnitElement.getType('C')!.fields[0];
+    var x = _resultUnitElement.getClass('C')!.fields[0];
     expect(x.name, 'x');
     _assertTypeStr(x.type, 'String');
-    var y = _resultUnitElement.getType('C')!.fields[1];
+    var y = _resultUnitElement.getClass('C')!.fields[1];
     expect(y.name, 'y');
     _assertTypeStr(y.type, 'Map<String, Map<String, String>>');
   }
@@ -3019,7 +3019,7 @@ class C {
     var x = _resultUnitElement.topLevelVariables[0];
     expect(x.name, 'x');
     _assertTypeStr(x.type, 'String');
-    var y = _resultUnitElement.getType('C')!.fields[0];
+    var y = _resultUnitElement.getClass('C')!.fields[0];
     expect(y.name, 'y');
     _assertTypeStr(y.type, 'String');
   }
@@ -3307,7 +3307,7 @@ class C {
       (int i) => {i: b};
 }
 ''');
-    var f = _resultUnitElement.getType('C')!.fields[0];
+    var f = _resultUnitElement.getClass('C')!.fields[0];
     _assertTypeStr(f.type, 'Map<int, bool> Function(int) Function(bool)');
   }
 
@@ -3317,7 +3317,7 @@ class C {
   static final f = (bool b) => b;
 }
 ''');
-    var f = _resultUnitElement.getType('C')!.fields[0];
+    var f = _resultUnitElement.getClass('C')!.fields[0];
     _assertTypeStr(f.type, 'bool Function(bool)');
   }
 
@@ -3327,7 +3327,7 @@ class C {
   static final f = (bool b) => 1;
 }
 ''');
-    var f = _resultUnitElement.getType('C')!.fields[0];
+    var f = _resultUnitElement.getClass('C')!.fields[0];
     _assertTypeStr(f.type, 'int Function(bool)');
   }